feat(page-path-truncation): relocate formatTruncatedPagePath to client/util
Move formatTruncatedPagePath (+ spec) from features/search/client/utils/
to client/util/ so search results list can reuse it, with no behavior
change. Update the sole existing consumer's import path.
