+++ apps/app/src/client/util/format-truncated-page-path.ts
